refactor: split layout/page.go into write and read halves
Move write-path functions to page_write.go and read-path functions to
page_read.go; split matching tests into page_write_test.go and
page_read_test.go; relocate TestTable_Pop_ArrayConsistency to table_test.go.
